requirements.txt
Django==2.2.10
pytz==2021.1
sqlparse==0.4.1
djangorestframework
django-debug-toolbar
djangorestframework-simplejwt
django-rest-swagger